backend/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "sqlacodegen>=4.0.1",
    "sqlalchemy>=2.0.47",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
    "langchain-community>=0.4.1",
    "dashscope>=1.25.13",
    "langchain>=1.2.10",
    "langgraph>=1.0.10",
    "langchain-deepseek>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "pymilvus>=2.6.9",
    "langchain-milvus>=0.3.3",
    "zhipuai>=2.1.5.20250825",
    "langsmith>=0.7.11",
    "tavily-python>=0.7.22",
]

[dependency-groups]
dev = [
    "sqlacodegen>=4.0.1",
]
